+++ packages/native-ui/src/components/List/ListHeading/styled-components/SupportingText.ts
@@ -1,0 +1,13 @@
+import { styled, Text } from '@gluestack-ui/themed';
+
+const ListHeadingSupportingText = styled(
+  Text,
+  {},
+  {
+    componentName: 'ListHeadingSupportingText',
+    descendantStyle: [],
+    ancestorStyle: ['_supportingText'],
+  }
+);
+
+export default ListHeadingSupportingText;