+++ packages/native-ui/src/components/List/ListItem/ListItem.props.ts
@@ -1,0 +1,14 @@
+import type { ComponentProps } from 'react';
+import type { Root } from './styled-components';
+
+interface ListItemProps extends ComponentProps<typeof Root> {
+  text: string;
+  supportingText?: string;
+  leadingContent?: React.ReactNode;
+  trailingContent?: React.ReactNode;
+  divider?: boolean;
+  loading?: boolean;
+  disabled?: boolean;
+}
+
+export default ListItemProps;