+++ packages/native-ui/src/components/List/ListItem/styled-components/Content.ts
@@ -1,0 +1,18 @@
+import { styled, VStack } from '@gluestack-ui/themed';
+
+const Content = styled(
+  VStack,
+  {
+    props: {
+      gap: '$1',
+      flex: 1,
+    },
+  },
+  {
+    componentName: 'ListItemContent',
+    descendantStyle: ['_text', '_supportingText'],
+    ancestorStyle: ['_content'],
+  }
+);
+
+export default Content;