+++ packages/native-ui/src/components/List/ListItem/styled-components/LeadingContent.ts
@@ -1,0 +1,13 @@
+import { styled, View } from '@gluestack-ui/themed';
+
+const LeadingContent = styled(
+  View,
+  {},
+  {
+    componentName: 'ListItemLeadingContent',
+    descendantStyle: ['_icon'],
+    ancestorStyle: ['_leadingContent'],
+  }
+);
+
+export default LeadingContent;