+++ packages/native-ui/src/components/List/ListItem/styled-components/SupportingText.ts
@@ -1,0 +1,13 @@
+import { styled, Text } from '@gluestack-ui/themed';
+
+const ListItemSupportingText = styled(
+  Text,
+  {},
+  {
+    componentName: 'ListItemSupportingText',
+    descendantStyle: [],
+    ancestorStyle: ['_supportingText'],
+  }
+);
+
+export default ListItemSupportingText;