+++ packages/native-ui/src/components/List/ListItem/styled-components/Text.ts
@@ -1,0 +1,13 @@
+import { styled, Text } from '@gluestack-ui/themed';
+
+const ListItemText = styled(
+  Text,
+  {},
+  {
+    componentName: 'ListItemText',
+    descendantStyle: [],
+    ancestorStyle: ['_text'],
+  }
+);
+
+export default ListItemText;