+++ packages/native-ui/src/components/List/ListItem/styled-components/TrailingContent.ts
@@ -1,0 +1,13 @@
+import { styled, View } from '@gluestack-ui/themed';
+
+const TrailingContent = styled(
+  View,
+  {},
+  {
+    componentName: 'ListItemTrailingContent',
+    descendantStyle: ['_icon'],
+    ancestorStyle: ['_trailingContent'],
+  }
+);
+
+export default TrailingContent;